requirements.txt
autoencoder==0.0.8
bcc==0.1.10
elasticsearch==8.12.1
kafka_python==2.0.2
mysql_connector_repackaged==0.3.1
numpy==1.26.4
pandas==2.2.1
Requests==2.31.0
scikit_learn==1.4.1.post1
tensorflow==2.15.0
tensorflow_cpu_aws==2.15.0
